% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdsky_plot_3.0.0.900.R
\name{bdsky_plot}
\alias{bdsky_plot}
\title{Plot Birth_Death Skyline Plots}
\usage{
bdsky_plot(logs = NULL, burninpercent = 10, recent = NULL,
  gridSize = 20, RepNumb = "R0.s",
  bUninfectiousRate = "becomeUninfectiousRate.s",
  sProportion = "samplingProportion.s")
}
\arguments{
\item{logs}{(character) The path to the log file(s) or the path to the file with
".txt" extension that stores the path to all log files to be
analyzed or NULL to select (default: NULL)}

\item{burninpercent}{(interger) The percentage of samples that should be ignored
from the log file}

\item{recent}{(interger) Date of the most recent sample (for plotting from past
to present)}

\item{gridSize}{(interger)}

\item{RepNumb}{(character) Name of the R0 parameter (default: "R0.s")}

\item{bUninfectiousRate}{(character) Name of the bacomeUninfectious parameter
(default: "becomeUninfectiousRate.s")}

\item{sProportion}{(character) Name of the samplingProportion parameter
(default: "samplingProportion.s")}
}
\value{
A PDF with the plots
}
\description{
This function plots the epidemiological parameters of a BDSKY analysis with
their HPD intervals.
}
\author{
Denise Kuehnert (denise.kuehnert@gmail.com)

Carlo Pacioni (carlo.pacioni@gmail.com)
}

