% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Prints the text 'Hello, world!'}
\usage{
hello()
}
\value{
nothing
}
\description{
Prints the text 'Hello, world!'
}
\author{
Default RStudio code, modified by Richel Bilderbeek
}

